requirements.txt
jax
matplotlib
dm-haiku